refactor: modularize loading logic
